Introduce TestVisibility::Context that now handles all the domain model creation and context propagation
