+++ sig/datadog/ci/test_visibility/store/global.rbs
@@ -1,7 +1,7 @@
 module Datadog
   module CI
     module TestVisibility
-      module Context
+      module Store
         class Global
           @mutex: Thread::Mutex
