+++ spec/datadog/ci/test_visibility/store/local_spec.rb
@@ -1,4 +1,8 @@
-RSpec.describe Datadog::CI::TestVisibility::Context::Local do
+# frozen_string_literal: true
+
+require_relative "../../../../../lib/datadog/ci/test_visibility/store/local"
+
+RSpec.describe Datadog::CI::TestVisibility::Store::Local do
   subject { described_class.new }
 
   let(:tracer_span) { double(Datadog::Tracing::SpanOperation, get_tag: "my test") }
